// src/tasks/type/task.type.ts
export enum Status {
  LOCKED = 'LOCKED', // TODO
  CLAIMABLE = 'CLAIMABLE',
  COMPLETED = 'COMPLETED',
  UNCOMPLETED = 'UNCOMPLETED',
}

export enum Type {
  ONE_TIME = 'ONE_TIME',
  FIXED = 'FIXED',
  DAILY = 'DAILY',
  PARTNER = 'PARTNER',
}

export enum Group {
  HOPIUM = 'HOPIUM',
  LUCKY = 'LUCKY',
}

export enum Prize {
  HOPIUM = 'HOPIUM',
  TICKET = 'TICKET',
}

export enum Condition {
  DAILY_CHECK_IN = 'DAILY_CHECK_IN',
  INVITE_FRIEND = 'INVITE_FRIEND',
  JOIN_TELEGRAM_GROUP = 'JOIN_TELEGRAM_GROUP',
  SUBSCRIBE_TELEGRAM_CHANNEL = 'SUBSCRIBE_TELEGRAM_CHANNEL',
  SUBSCRIBE_TWITTER = 'SUBSCRIBE_TWITTER',
  CONNECT_TON_WALLET = 'CONNECT_TON_WALLET',
  WIN_PVP_BATTLE = 'WIN_PVP_BATTLE',
  AFFILIATE_CLICK = 'AFFILIATE_CLICK',
  ONE_FORECAST = 'ONE_FORECAST',
  THREE_FORECAST = 'THREE_FORECAST',
  FIVE_FORECAST = 'FIVE_FORECAST',
}